util: Move TimeRange into its own time.go
